refactor: character sprite should derive textures from its state
