+++ apps/game/client/actor/actor.tsx
@@ -1,22 +1,10 @@
-import type { Spritesheet } from "pixi.js";
 import { Container, Text } from "pixi.js";
 import { Pixi } from "@mp/solid-pixi";
-import {
-  createContext,
-  createEffect,
-  createMemo,
-  Show,
-  useContext,
-} from "solid-js";
-import { assert } from "@mp/std";
+import { createEffect, createMemo, Show } from "solid-js";
 import type { TiledResource } from "../../shared/area/tiled-resource";
 import type { Actor } from "../../server/traits/actor";
 import { createTintFilter } from "../tint-filter";
 import { createCharacterSprite } from "./character-sprite";
-import type {
-  CharacterModelId,
-  CharacterSpriteState,
-} from "./character-sprite-state";
 import { loopedCharacterSpriteStates } from "./character-sprite-state";
 import { deriveCharacterSpriteState } from "./character-sprite-state-for-actor";
 
@@ -30,16 +18,12 @@ export function Actor(props: {
   );
 
   const container = new Container();
-  const allSpriteshets = useContext(CharacterSpritesheetContext);
 
   const state = createMemo(() => deriveCharacterSpriteState(props.actor));
 
   const sprite = createCharacterSprite(
+    () => state(),
     () => props.actor.dir,
-    () =>
-      assert(
-        allSpriteshets.get("adventurer" as CharacterModelId)?.get(state()),
-      ),
   );
 
   createEffect(() => {
@@ -69,17 +53,3 @@ export function Actor(props: {
     </Show>
   );
 }
-
-export const CharacterSpritesheetContext = createContext(
-  new Proxy(
-    {} as ReadonlyMap<
-      CharacterModelId,
-      ReadonlyMap<CharacterSpriteState, Spritesheet>
-    >,
-    {
-      get() {
-        throw new Error("CharacterSpritesheetContext is not initialized");
-      },
-    },
-  ),
-);
