+++ apps/game/client/actor/character-sprite.tsx
@@ -1,18 +1,30 @@
 import type { AnimatedSpriteFrames, Spritesheet } from "pixi.js";
 import { AnimatedSprite } from "pixi.js";
 import type { Accessor } from "solid-js";
-import { createEffect, createMemo } from "solid-js";
+import { createContext, createEffect, createMemo, useContext } from "solid-js";
 import {
   cardinalDirectionAngles,
   cardinalDirections,
   nearestCardinalDirection,
   type CardinalDirection,
 } from "@mp/math";
+import { assert } from "@mp/std";
+import type {
+  CharacterModelId,
+  CharacterSpriteState,
+} from "./character-sprite-state";
 
 export function createCharacterSprite(
+  state: Accessor<CharacterSpriteState>,
   desiredDirection: Accessor<CardinalDirection>,
-  spritesheet: Accessor<Spritesheet>,
 ): AnimatedSprite {
+  const allSpriteshets = useContext(CharacterSpritesheetContext);
+  const spritesheet = createMemo(
+    (): Spritesheet =>
+      assert(
+        allSpriteshets.get("adventurer" as CharacterModelId)?.get(state()),
+      ),
+  );
   const direction = createMemo(() =>
     spritesheetCompatibleDirection(desiredDirection(), spritesheet()),
   );
@@ -20,7 +32,7 @@ export function createCharacterSprite(
     const textures = spritesheet().animations[direction()];
     return textures.map((texture) => ({
       texture,
-      time: 100,
+      time: 100, // TODO should come as metadata from the spritesheet
     }));
   });
 
@@ -59,3 +71,17 @@ function spritesheetCompatibleDirection(
   const desiredAngle = cardinalDirectionAngles[desiredDirection];
   return nearestCardinalDirection(desiredAngle, availableDirections);
 }
+
+export const CharacterSpritesheetContext = createContext(
+  new Proxy(
+    {} as ReadonlyMap<
+      CharacterModelId,
+      ReadonlyMap<CharacterSpriteState, Spritesheet>
+    >,
+    {
+      get() {
+        throw new Error("CharacterSpritesheetContext is not initialized");
+      },
+    },
+  ),
+);
